Introduce DataLoader
As of now there is a single LoaderInterface::load() method which takes a file as a first parameter. However in some cases you want to be able to provide the data right away instead of having to use a file. For this purpose, the proposed changes are:

- The file Parsing has been moved out of FixtureBuilder and Introduce DataLoaderInterface::loadData() which takes an array of data as a first parameter.
- To avoid any confusion with DataLoaderInsterface, LoaderInterface::load() has been renamed FileLoaderInterface::loadFile().
