+++ fixtures/Nelmio/Alice/FixtureBuilder/FakeFixtureBuilder.php
@@ -1,0 +1,29 @@
+<?php
+
+/*
+ * This file is part of the Alice package.
+ *
+ * (c) Nelmio <hello@nelm.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Nelmio\Alice\FixtureBuilder;
+
+use Nelmio\Alice\FixtureBuilderInterface;
+use Nelmio\Alice\FixtureSet;
+use Nelmio\Alice\NotCallableTrait;
+
+class FakeFixtureBuilder implements FixtureBuilderInterface
+{
+    use NotCallableTrait;
+
+    /**
+     * @inheritdoc
+     */
+    public function build(array $data, array $parameters = [], array $objects = []): FixtureSet
+    {
+        $this->__call();
+    }
+}
