+++ tests/Nelmio/Alice/FixtureBuilder/SimpleBuilderTest.php
@@ -13,11 +13,11 @@ namespace Nelmio\Alice\FixtureBuilder;
 
 use Nelmio\Alice\FixtureBag;
 use Nelmio\Alice\FixtureBuilder\Denormalizer\FakeDenormalizer;
-use Nelmio\Alice\FixtureBuilder\Parser\FakeParser;
 use Nelmio\Alice\FixtureBuilderInterface;
 use Nelmio\Alice\FixtureSet;
 use Nelmio\Alice\ObjectBag;
 use Nelmio\Alice\ParameterBag;
+use Nelmio\Alice\ParserInterface;
 use Prophecy\Argument;
 
 /**
@@ -35,65 +35,51 @@ class SimpleBuilderTest extends \PHPUnit_Framework_TestCase
      */
     public function testIsNotClonable()
     {
-        $builder = new SimpleBuilder(new FakeParser(), new FakeDenormalizer());
+        $builder = new SimpleBuilder(new FakeDenormalizer());
         clone $builder;
     }
 
     public function testBuildSet()
     {
-        $file = 'dummy.yml';
+        $data = ['dummy' => new \stdClass()];
         $injectedParameters = ['foo' => 'bar'];
         $injectedObjects = ['std' => new \stdClass()];
 
-        $parseData = ['dummy' => new \stdClass()];
         $loadedParameters = new ParameterBag(['rab' => 'oof']);
         $loadedFixtures = new FixtureBag();
         $set = new BareFixtureSet($loadedParameters, $loadedFixtures);
 
         $expected = new FixtureSet($loadedParameters, new ParameterBag($injectedParameters), $loadedFixtures, new ObjectBag($injectedObjects));
 
-        $parserProphecy = $this->prophesize(ParserInterface::class);
-        $parserProphecy->parse($file)->willReturn($parseData);
-        /** @var ParserInterface $parser */
-        $parser = $parserProphecy->reveal();
-
         $denormalizerProphecy = $this->prophesize(DenormalizerInterface::class);
-        $denormalizerProphecy->denormalize($parseData)->willReturn($set);
+        $denormalizerProphecy->denormalize($data)->willReturn($set);
         /** @var DenormalizerInterface $denormalizer */
         $denormalizer = $denormalizerProphecy->reveal();
 
-        $builder = new SimpleBuilder($parser, $denormalizer);
-        $actual = $builder->build($file, $injectedParameters, $injectedObjects);
+        $builder = new SimpleBuilder($denormalizer);
+        $actual = $builder->build($data, $injectedParameters, $injectedObjects);
 
         $this->assertEquals($expected, $actual);
 
-        $parserProphecy->parse(Argument::any())->shouldHaveBeenCalledTimes(1);
         $denormalizerProphecy->denormalize(Argument::any())->shouldHaveBeenCalledTimes(1);
     }
 
     public function testBuildSetWithoutInjectingParametersOrObjects()
     {
-        $file = 'dummy.yml';
-
-        $parseData = ['dummy' => new \stdClass()];
+        $data = ['dummy' => new \stdClass()];
         $loadedParameters = new ParameterBag(['rab' => 'oof']);
         $loadedFixtures = new FixtureBag();
         $set = new BareFixtureSet($loadedParameters, $loadedFixtures);
 
         $expected = new FixtureSet($loadedParameters, new ParameterBag(), $loadedFixtures, new ObjectBag());
 
-        $parserProphecy = $this->prophesize(ParserInterface::class);
-        $parserProphecy->parse($file)->willReturn($parseData);
-        /** @var ParserInterface $parser */
-        $parser = $parserProphecy->reveal();
-
         $denormalizerProphecy = $this->prophesize(DenormalizerInterface::class);
-        $denormalizerProphecy->denormalize($parseData)->willReturn($set);
+        $denormalizerProphecy->denormalize($data)->willReturn($set);
         /** @var DenormalizerInterface $denormalizer */
         $denormalizer = $denormalizerProphecy->reveal();
 
-        $builder = new SimpleBuilder($parser, $denormalizer);
-        $actual = $builder->build($file);
+        $builder = new SimpleBuilder($denormalizer);
+        $actual = $builder->build($data);
 
         $this->assertEquals($expected, $actual);
     }
