+++ tests/Nelmio/Alice/Parser/FakeParser.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the Alice package.
+ *  
+ * (c) Nelmio <hello@nelm.io>
+ *  
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Nelmio\Alice\Parser;
+
+use Nelmio\Alice\ParserInterface;
+
+class FakeParser implements ParserInterface
+{
+    /**
+     * @inheritdoc
+     */
+    public function parse(string $file): array
+    {
+        throw new \BadMethodCallException();
+    }
+}
