+++ tests/Nelmio/Alice/Parser/FileListProviderTrait.php
@@ -1,0 +1,30 @@
+<?php
+
+/*
+ * This file is part of the Alice package.
+ *
+ * (c) Nelmio <hello@nelm.io>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Nelmio\Alice\Parser;
+
+trait FileListProviderTrait
+{
+    public function providePhpList()
+    {
+        return FilesReference::getPhpList();
+    }
+
+    public function provideYamlList()
+    {
+        return FilesReference::getYamlList();
+    }
+
+    public function provideUnsupportedList()
+    {
+        return FilesReference::getUnsupportedList();
+    }
+}
